rename files and change render big picture
