+++ js/main.js
@@ -1,4 +1,4 @@
-import { renderPhotos } from './renderPhotos.js';
+import { renderPhotos } from './render-photos.js';
 import { createUsersPhotos } from './data.js';
 
 const data = createUsersPhotos();
