+++ js/render-big-picture.js
@@ -1,9 +1,14 @@
 const commentTemplate = document.querySelector('.social__comment');
 const commentsList = document.querySelector('.social__comments');
-commentsList.textContent = '';
+const bigPicture = document.querySelector('.big-picture');
+const countLikes = bigPicture.querySelector('.likes-count');
+const photoImage = bigPicture.querySelector('.big-picture__img img');
+const photoTitle = bigPicture.querySelector('.social__caption');
 
 
 const renderComments = (comments) => {
+  commentsList.innerHTML = '';
+
   comments.forEach((comment) => {
     const newComment = commentTemplate.cloneNode(true);
     const userPhoto = newComment.querySelector('img');
@@ -17,4 +22,14 @@ const renderComments = (comments) => {
   });
 };
 
-export {renderComments};
+
+const renderBigPicture = (url, description, comments, likes) => {
+  photoImage.src = url;
+  photoImage.alt = description;
+  countLikes.textContent = likes;
+  photoTitle.textContent = description;
+
+  renderComments(comments);
+};
+
+export {renderBigPicture};
